Account for different versions of Test Kitchen
Support for synchronizing +converge+ was introduced in
[Test Kitchen v1.19.0](https://github.com/test-kitchen/test-kitchen/blob/v1.19.0/lib/kitchen/driver/base.rb#L97)
so at least that version is required to serialize all Terraform commands.

Since support for Ruby v2.2 was dropped in
[Test Kitchen v1.17.0](https://github.com/test-kitchen/test-kitchen/blob/v1.17.0/test-kitchen.gemspec#L24),
we must also drop support for Ruby v2.2 in order to depend on Test
Kitchen v1.19.0.

Since there are no breaking changes between Test Kitchen v1.16.0 and
v1.20.0, we can depend on the latest version.