Normalize format, escaping of Terraform options (#296)
* Normalize format, escaping of Terraform options

Escaping options which accept arguments of HCL is unreliable. This
commit removes escaping from -backend-config and -var, and applies
single quotes consistently to complex option arguments.

* Double quote Terraform path option arguments

* Escape -backend-config, -var without quotes

* Remove length limit of RSpec output

This change ensures failure output will not be truncated.

* Double quote -backend-config, -var; drop escaping

Windows requires any argument with space to be double quoted. Using
double quotes for all arguments and deferring to the user to escape HCL
input appropriately seems to be the only functional solution.

* Update version to 4.0.5

* Update bundles

* Update Change Log